fix conflict

+++ script.js
@@ -1,5 +1,7 @@
 function resetCart() {
-  cartCount = 0;
-  localStorage.setItem("cartCount", cartCount);
-  document.getElementById("cart-count").textContent = cartCount;
+  if (confirm("カートを空にしますか？")) {
+    cartCount = 0;
+    localStorage.setItem("cartCount", cartCount);
+    document.getElementById("cart-count").textContent = cartCount;
+  }
 }
